+++ pkgs/async/lib/src/result/result.dart
@@ -57,6 +57,7 @@ abstract class Result<T> {
     ReleaseStreamTransformer<Object>(),
   );
 
+
   /// Creates a `Result` with the result of calling [computation].
   ///
   /// This generates either a [ValueResult] with the value returned by
@@ -219,6 +220,13 @@ abstract class Result<T> {
   ///
   /// Calls the sink's `add` or `addError` method as appropriate.
   void addTo(EventSink<T> sink);
+
+  /// The value of this result, or throws the error if it is an error result.
+  ///
+  /// If this is a [ValueResult], returns its value.
+  /// If this is an [ErrorResult], throws its error with its stack trace
+  /// using [Error.throwWithStackTrace].
+  T get valueOrRethrow;
 
   /// A future that has been completed with this result as a value or an error.
   Future<T> get asFuture;